Add reusable code for SimpleSerialize for cbor-serialized structs
There are multiple structs (and more to be added in subsequent
commits) that are used as values in TreeMap, and serialized using
serde_cbor. This commit adds a reusable implementation by defining two
new marker traits 'CborSerialize' and 'CborDeserialize'. Any struct
that implements these two traits will automatically get the
'SimpleSerialize' trait implemented for it.
